feat add different animals classes extends AbstractAnimal

+++ src/lessons/java/animals/AbstractAnimal.java
@@ -4,9 +4,9 @@ public abstract class AbstractAnimal {
 
     private String name;
     private int age;
-    private int weight;
+    private float weight;
 
-    public AbstractAnimal(String name, int age, int weight) {
+    public AbstractAnimal(String name, int age, float weight) {
         this.name = name;
         this.age = age;
         this.weight = weight;
@@ -34,12 +34,12 @@ public abstract class AbstractAnimal {
         this.age = age;
     }
 
-    public int getWeight() {
+    public float getWeight() {
 
         return this.weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(float weight) {
 
         this.weight = weight;
     }
